server/requirements.txt
pyserial>=3.5
flask>=3.0
openai>=1.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
python-dotenv>=1.0
